+++ app/Http/Controllers/ContactController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Storage;
+use Gate;
 
 class ContactController extends Controller
 {
@@ -32,34 +33,38 @@ class ContactController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
-        if ($request->gender == 'pilih'){
-            return back()->withErrors(['gender' => ['Silahkan pilih jenis kelamin']])->withInput();
+        if((Gate::allows('isAdmin') && Gate::allows('isClient') && Gate::allows('isMarketingCreate')) || (Gate::allows('isMarketing') && Gate::allows('isClient') && Gate::allows('isMarketingCreate'))){
+            if ($request->gender == 'pilih'){
+                return back()->withErrors(['gender' => ['Silahkan pilih jenis kelamin']])->withInput();
+            }
+            $request->request->add(['client_id' => $request->client_id, 'user_id' => auth()->user()->id]);
+            $rules = [
+                'user_id' => 'required',
+                'client_id' => 'required',
+                'gender' => 'required',
+                'name' => 'required|max:255',
+                'photo' => 'image|file|max:1024'
+            ];
+            if($request->email){
+                $rules['email'] = 'email:dns|unique:contacts';
+            }
+            if($request->phone){
+                $rules['phone'] = 'min:10|max:15|unique:contacts';
+            }
+    
+            $validateData = $request->validate($rules);
+            $validateData['position'] = $request->position;
+    
+            if($request->file('photo')){
+                $validateData['photo'] = $request->file('photo')->store('contact-images');
+            }
+    
+            Contact::create($validateData);
+    
+            return redirect('/marketing/clients/'. $request->client_id)->with('success','Kontak baru dengan nama '. $request->name . ' berhasil ditambahkan');
+        } else {
+            abort(403);
         }
-        $request->request->add(['client_id' => $request->client_id, 'user_id' => auth()->user()->id]);
-        $rules = [
-            'user_id' => 'required',
-            'client_id' => 'required',
-            'gender' => 'required',
-            'name' => 'required|max:255',
-            'photo' => 'image|file|max:1024'
-        ];
-        if($request->email){
-            $rules['email'] = 'email:dns|unique:contacts';
-        }
-        if($request->phone){
-            $rules['phone'] = 'min:10|max:15|unique:contacts';
-        }
-
-        $validateData = $request->validate($rules);
-        $validateData['position'] = $request->position;
-
-        if($request->file('photo')){
-            $validateData['photo'] = $request->file('photo')->store('contact-images');
-        }
-
-        Contact::create($validateData);
-
-        return redirect('/media/clients/'. $request->client_id)->with('success','Kontak baru dengan nama '. $request->name . ' berhasil ditambahkan');
     }
 
     /**
@@ -75,10 +80,14 @@ class ContactController extends Controller
      */
     public function edit(Contact $contact): Response
     {
-        return response()->view('contacts.edit', [
-            'contact' => $contact,
-            'title' => 'Edit Kontak Person'
-        ]);
+        if((Gate::allows('isAdmin') && Gate::allows('isClient') && Gate::allows('isMarketingEdit')) || (Gate::allows('isMarketing') && Gate::allows('isClient') && Gate::allows('isMarketingEdit'))){
+            return response()->view('contacts.edit', [
+                'contact' => $contact,
+                'title' => 'Edit Kontak Person'
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -86,45 +95,49 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact): RedirectResponse
     {
-        $request->request->add(['client_id' => $request->client_id, 'user_id' => auth()->user()->id]);
-        $rules = [
-            'user_id' => 'required',
-            'client_id' => 'required',
-            'gender' => 'required',
-            'name' => 'required|max:255',
-            'photo' => 'image|file|max:1024'
-        ];
-
-        if($request->email){
-            if($request->email != $contact->email){
-                $rules['email'] = 'email:dns|unique:contacts';
-            } 
-        }elseif($contact->email){
-            $rules['email'] = '';
-        }
-
-        if($request->phone){
-            if($request->phone != $contact->phone){
-                $rules['phone'] = 'min:10|max:15|unique:contacts';
+        if((Gate::allows('isAdmin') && Gate::allows('isClient') && Gate::allows('isMarketingEdit')) || (Gate::allows('isMarketing') && Gate::allows('isClient') && Gate::allows('isMarketingEdit'))){
+            $request->request->add(['client_id' => $request->client_id, 'user_id' => auth()->user()->id]);
+            $rules = [
+                'user_id' => 'required',
+                'client_id' => 'required',
+                'gender' => 'required',
+                'name' => 'required|max:255',
+                'photo' => 'image|file|max:1024'
+            ];
+    
+            if($request->email){
+                if($request->email != $contact->email){
+                    $rules['email'] = 'email:dns|unique:contacts';
+                } 
+            }elseif($contact->email){
+                $rules['email'] = '';
             }
-        }elseif($contact->phone){
-            $rules['phone'] = '';
-        }
-
-        $validateData = $request->validate($rules);
-        $validateData['position'] = $request->position;
-
-        if($request->file('photo')){
-            if($request->oldPhoto){
-                Storage::delete($request->oldPhoto);
+    
+            if($request->phone){
+                if($request->phone != $contact->phone){
+                    $rules['phone'] = 'min:10|max:15|unique:contacts';
+                }
+            }elseif($contact->phone){
+                $rules['phone'] = '';
             }
-            $validateData['photo'] = $request->file('photo')->store('contact-images');
+    
+            $validateData = $request->validate($rules);
+            $validateData['position'] = $request->position;
+    
+            if($request->file('photo')){
+                if($request->oldPhoto){
+                    Storage::delete($request->oldPhoto);
+                }
+                $validateData['photo'] = $request->file('photo')->store('contact-images');
+            }
+    
+            Contact::where('id', $contact->id)
+                    ->update($validateData);
+    
+            return redirect('/marketing/clients/'. $request->client_id)->with('success','Kontak person dengan nama '. $request->name . ' berhasil dirubah');
+        } else {
+            abort(403);
         }
-
-        Contact::where('id', $contact->id)
-                ->update($validateData);
-
-        return redirect('/media/clients/'. $request->client_id)->with('success','Kontak person dengan nama '. $request->name . ' berhasil dirubah');
     }
 
     /**
@@ -132,13 +145,17 @@ class ContactController extends Controller
      */
     public function destroy(Contact $contact): RedirectResponse
     {
-        if($contact->photo){
-            Storage::delete($contact->photo);
+        if((Gate::allows('isAdmin') && Gate::allows('isClient') && Gate::allows('isMarketingDelete')) || (Gate::allows('isMarketing') && Gate::allows('isClient') && Gate::allows('isMarketingDelete'))){
+            if($contact->photo){
+                Storage::delete($contact->photo);
+            }
+    
+            Contact::destroy($contact->id);
+    
+            return redirect('/marketing/clients/'. $contact->client_id)->with('success','Kontak person dengan nama' . $contact->name . ' berhasil dihapus');
+        } else {
+            abort(403);
         }
-
-        Contact::destroy($contact->id);
-
-        return redirect('/media/clients/'. $contact->client_id)->with('success','Kontak person dengan nama' . $contact->name . ' berhasil dihapus');
     }
 
     /**
